requirements.txt

Werkzeug==2.0.1
Pillow==10.3.0  # Updated version
numpy==1.26.4  # Updated version
torch>=2.0.0
torchvision
flask
gunicorn